requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
blis==0.2.4
certifi==2017.11.5
chardet==3.0.4
ChatterBot==1.0.5
chatterbot-corpus==1.2.0
cymem==2.0.3
feedparser==5.2.1
idna==2.6
mathparse==0.1.2
multidict==4.7.5
murmurhash==1.0.2
nltk==3.4.5
numpy==1.18.1
Pint==0.11
plac==0.9.6
preshed==2.0.1
prettytable==0.7.2
pymongo==3.10.1
python-dateutil==2.7.5
pytz==2019.3
PyYAML==3.13
requests==2.18.4
six==1.11.0
slackclient==2.5.0
spacy==2.1.9
SQLAlchemy==1.3.15
srsly==1.0.2
thinc==7.0.8
tqdm==4.43.0
urllib3==1.22
wasabi==0.6.0
websocket-client==0.44.0
yarl==1.4.2
